% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowAnalysis.R
\name{fhAnalyze}
\alias{fhAnalyze}
\title{fhAnalyze}
\usage{
fhAnalyze(fh)
}
\arguments{
\item{fh}{a \code{FlowHist} object}
}
\value{
a \code{FlowHist} object with the analysis (nls, counts, cv,
  RCS) slots filled.
}
\description{
Complete non-linear regression analysis of FlowHist histogram data
}
\details{
Completes the NLS analysis, and calculates the modelled events and CVs
for the result.
}
\examples{
library(flowPloidyData)
fh1 <- FlowHist(file = flowPloidyFiles[1], channel = "FL3.INT.LIN")
fh1 <- fhAnalyze(fh1)
}
\author{
Tyler Smith
}
\seealso{
\code{\link{FlowHist}}
}

