% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowAnalysis.R
\name{fhDoCounts}
\alias{fhDoCounts}
\title{Calculcate \code{\link{FlowHist}} nuclei counts}
\usage{
fhDoCounts(fh)
}
\arguments{
\item{fh}{a \code{\link{FlowHist}} object}
}
\value{
The updated \code{\link{FlowHist}} object with the \code{counts}
  slot updated.
}
\description{
Uses the fitted NLS model for a \code{\link{FlowHist}} object to
calculate the cell counts for the G1 peak, G2 peak, and S-phase model
components. The actual values are generated by numerical integration of
the model components.
}
\seealso{
\code{\link{integrate}}, \code{\link{fhDoCV}}
  \code{\link{fhDoNLS}}, \code{\link{fhDoRCS}}.
}
\author{
Tyler Smith
}
\keyword{internal}
