% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowAnalysis.R
\name{fhDoNLS}
\alias{fhDoNLS}
\title{Fit the NLS model for a \code{\link{FlowHist}} model}
\usage{
fhDoNLS(fh)
}
\arguments{
\item{fh}{a \code{\link{FlowHist}} object}
}
\value{
The \code{\link{FlowHist}} object with the \code{NLS} slot
  updated to include the results of the analysis
}
\description{
Constructs the call to \code{\link{nlsLM}} for the
\code{\link{FlowHist}} object.
}
\seealso{
\code{\link{nlsLM}}, \code{\link{fhDoCV}},
  \code{\link{fhDoRCS}}, \code{\link{fhDoCounts}}
}
\author{
Tyler Smith
}
\keyword{internal}
