% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowAnalysis.R
\name{fhDoRCS}
\alias{fhDoRCS}
\title{Calculcate the Residual Chi-Square for a \code{\link{FlowHist}} object}
\usage{
fhDoRCS(fh)
}
\arguments{
\item{fh}{a \code{\link{FlowHist}} object}
}
\value{
The updated \code{\link{FlowHist}} object.
}
\description{
Calculcate the Residual Chi-Square value for a \code{\link{FlowHist}}
model fit.
}
\details{
This is only a rough guideline, and needs to be interpreted with some
caution. The RCS is heavily influenced by the highest channels, where
the expected value is close to zero. Consequently, observing 2 or 3
stray events in one of these channels, where the expected value may be <
0.02, produces a higher value of (obs - exp)^2 / exp than larger
absolute differences in the main region of the histogram.
                                                                    
Rabinovitch 1994:                                                   
                                                                    
The x2 is affected by a large number of variables, not all related  
to goodness of the fit; these include the number of cells acquired  
in the histogram and the end points of the analysis region used     
within the histogram.
}
\seealso{
\code{\link{fhDoCV}}, \code{\link{fhDoNLS}},
  \code{\link{fhDoCounts}}
}
\author{
Tyler Smith
}
\keyword{internal}
