% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowHist.R
\name{fhMetadata}
\alias{fhMetadata}
\alias{viewFlowChannels}
\title{Display metadata present in an FCS file}
\usage{
viewFlowChannels(file, emptyValue = TRUE, truncate_max_range = TRUE)

fhMetadata(file, emptyValue = TRUE, truncate_max_range = TRUE)
}
\arguments{
\item{file}{character, the name of an FCS data file; or the name of a
FlowHist object.}

\item{emptyValue}{boolean, passed to \code{\link{read.FCS}},
needed to deal with unusual FCS file formats. Default is TRUE - if
your file loads without errors, then don't change this value}

\item{truncate_max_range}{boolean, passed to \code{\link{read.FCS}}.}
}
\value{
A list of
}
\description{
Displays a selection of potentially useful metadata fields from the FCS
data file, and a summary of the data channels. Useful for selecting the
the \code{channel} argument in \code{\link{FlowHist}}, finding a
starting value for \code{trimRaw} (when needed), and identifying the
parameter to use for the sample name.
}
\details{
The `top` column in the \code{Flow Channels} table displays the 99th
percentile value for that parameter. For flow cytometers that require
the \code{trimRaw} argument, this may be a reasonable starting place.
See \code{\link{FlowHist}}.
}
\examples{
library(flowPloidyData) 
fhMetadata(flowPloidyFiles()[1])

## retained for now, but viewFlowChannels will be removed from future
## relases: 
viewFlowChannels(flowPloidyFiles()[1])
}
\seealso{
\code{\link{FlowHist}}
}
\author{
Tyler Smith
}
