% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowHist.R
\name{findPeaks}
\alias{cleanPeaks}
\alias{findPeaks}
\title{findPeaks}
\usage{
findPeaks(fh, window, smooth = window/2)

cleanPeaks(fh, window)
}
\arguments{
\item{fh}{a \code{FlowHist} object}

\item{window}{an integer, the width of the moving window to use in
identifying local maxima via \code{caTools::runmax}}

\item{smooth}{an integer, the width of the moving window to use in
removing noise via \code{caTools::runmean}}
}
\value{
Returns a matrix with two columns:
\describe{
\item{mean}{the index position of each potential peak}
\item{height}{the height (intensity) of the peak at that index position}
}
}
\description{
Locate potential peaks in histogram data
}
\details{
Peaks are defined as local maxima in the vector of values, using a
moving window. Note that these are used in the context of finding
starting values - accuracy isn't important, we just need something
`close-enough' that the nls algorithm will be able to find the correct
value.

Utility functions for use internally by flowPloidy; not exported and
won't be visible to users. Usually invoked from within \code{FlowHist}.

Note that there is a trade-off between accuracy in detected peaks, and
avoiding noise. Increasing the value of \code{smooth} will reduce the
amount of 'noise' that is included in the peak list. However, increasing
smoothing shifts the location of the actual peaks. Most of the time the
default values provide an acceptable compromise, given we only need to
get 'close enough' for the NLS optimization to find the true parameter
values. If you'd like to explore this, the internal (unexported)
function \code{fhPeakPlot} may be useful.

\code{cleanPeaks} filters the output of \code{findPeaks} to:  
\itemize{

\item remove duplicates, ie., peaks with the same intensity that occur
within \code{window} positions of each other. Otherwise,
\code{findPeaks} will consider noisy peaks without a single highest
point to be multiple distinct peaks.

\item drop G2 peaks. In some cases the G2 peak for one sample will have
greater intensity than the G1 peak for another sample. We correct for
this by removing detected peaks with means close to twice that of other
peaks.

\item ignore noise, by removing peaks with \code{intensity} < 40. A
somewhat arbitrary value. It's tricky to deal with this issue when the
debris field is large.
}
}
\examples{
\dontrun{
set.seed(123)
test.dat <-(cumsum(runif(1000, min = -1)))
plot(test.dat, type = 'l')
test.peaks <- flowPloidy::findPeaks(test.dat, window = 20)
points(test.peaks, col = 'red', cex = 2)
}

}
\author{
Tyler Smith
}

