% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowHist.R
\name{plot.FlowHist}
\alias{plot.FlowHist}
\title{Plot histograms for FlowHist objects}
\usage{
\method{plot}{FlowHist}(x, init = FALSE, nls = TRUE, comps = TRUE,
  main = fhFile(x), ...)
}
\arguments{
\item{x}{a \code{\link{FlowHist}} object}

\item{init}{boolean; if TRUE, plot the regression model using the
initial parameter estimates over the raw data.}

\item{nls}{boolean; if TRUE, plot the fitted regression model over the
raw data (i.e., using the final parameter values)}

\item{comps}{boolean; if TRUE, plot the individual model components
over the raw data.}

\item{...}{additional arguments passed on to plot()}
}
\value{
Not applicable
}
\description{
Plot histograms for FlowHist objects
}
\author{
Tyler Smith
}

