% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFlowHist.R
\name{plotFH}
\alias{plotFH}
\title{Plot the raw data for a FlowHist object}
\usage{
plotFH(fh, main = fhFile(fh), ...)
}
\arguments{
\item{fh}{a \code{\link{FlowHist}} object}

\item{main}{character; the plot title. Defaults to the filename of the
\code{\link{FlowHist}} object.}

\item{...}{additional parameters passed to \code{\link{plot}}}
}
\value{
Not applicable, used for plotting
}
\description{
Creates a simple plot of the raw histogram data. Used as a utility for
other plotting functions, and perhaps useful for users who wish to
create their own plotting routines.
}
\examples{
library(flowPloidyData) 
fh1 <- FlowHist(file = flowPloidyFiles()[1], channel = "FL3.INT.LIN")
plotFH(fh1)
}
\author{
Tyler Smith
}
