% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowHist.R
\name{tabulateFlowHist}
\alias{tabulateFlowHist}
\title{exportFlowHist}
\usage{
tabulateFlowHist(fh, file = NULL)
}
\arguments{
\item{fh}{a FlowHist object, or a list of FlowHist objects.}

\item{file}{character, the name of the file to save data to}
}
\value{
a data frame
}
\description{
Extract analysis results from a FlowHist object
}
\details{
A convenience function for extracting the results of the NLS
  curve-fitting analysis on a FlowHist object.

If \code{fh} is a single FlowHist object, a data.frame with a single
row is returned. If \code{fh} is a list of \code{\link{FlowHist}} objects, a
row for each object will be added to the data.frame.

If a file name is provided, the data will be saved to that file.
}
\examples{
library(flowPloidyData) 
fh1 <- FlowHist(file = flowPloidyFiles[1], channel = "FL3.INT.LIN")
fh1 <- fhAnalyze(fh1)
tabulateFlowHist(fh1)
}
\author{
Tyler Smith
}

