% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowHist.R
\name{viewFlowChannels}
\alias{viewFlowChannels}
\title{viewFlowChannels}
\usage{
viewFlowChannels(file)
}
\arguments{
\item{file}{character, the name of a FCS data file}
}
\value{
A vector of column names from the FCS file.
}
\description{
Displays the column names present in an FCS file
}
\details{
A convenience function for viewing column names in an FCS, in order to
  select one for the \code{channel} argument in \code{FlowHist}.
}
\examples{
library(flowPloidyData) 
viewFlowChannels(flowPloidyFiles[1])
}
\author{
Tyler Smith
}
\seealso{
\code{FlowHist}
}

