% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{browseFlowHist}
\alias{browseFlowHist}
\title{browseFlowHist}
\usage{
browseFlowHist(flowList, debug = FALSE)
}
\arguments{
\item{flowList}{either a \code{\link{FlowHist}} object, or a list of
\code{\link{FlowHist}} objects}

\item{debug}{boolean, turns on debugging messages}
}
\value{
Returns the list of \code{\link{FlowHist}} objects, updated by any
  changes made in the GUI.
}
\description{
Visually assess histogram fits
}
\details{
Visually assess histogram fits
}
\examples{
library(flowPloidyData)
batch1 <- batchFlowHist(flowPloidyFiles, channel = "FL3.INT.LIN")
\dontrun{
batch1 <- browseFlowHist(batch1)
}
}
\author{
Tyler Smith
}

