% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gauss}
\alias{gauss}
\title{Gaussian model components}
\arguments{
\item{a1, a2, b1, b2}{area parameters}

\item{Ma, Mb}{curve mean parameter}

\item{Sa, Sb}{curve standard deviation parameter}

\item{xx}{vector of histogram intensities}
}
\value{
NA
}
\description{
Components for modeling Gaussian features in flow histograms
}
\details{
Typically the complete models will contain fA1 and fB2, which model the
G1 peaks of the sample and the standard. In most cases, they will also
contain fA2 and fB2, which model the G2 peaks. The G2 peaks are linked
to the G1 peaks, in that they require some of the parameters from the
G1 peaks as well (mean and standard deviation).
}
\author{
Tyler Smith
}

