% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowHist.R
\name{pickInit}
\alias{pickInit}
\title{Interactively select model starting values}
\usage{
pickInit(fh)
}
\arguments{
\item{fh}{A \code{flowHist} object}
}
\value{
\code{pickInit} returns the \code{flowHist} object with its
  initial value slot updated.

\code{pickPeaks} returns a matrix with each peak as a row, with the mean
(position) in the first column, and the height (intensity) in the second
column.
}
\description{
Prompts the user to select the peaks to use as initial
  values for non-linear regression on a plot of the histogram data.
}
\details{
The raw histogram data are plotted, and the user is prompted to
  select the peak positions to use as starting values in the NLS
  procedure. This is useful when the automated peak-finding algorithm
  fails to discriminate between overlapping peaks, or is confused by
  noise.

The normal use, \code{pickPeaks} is called from \code{pickInit}, rather
  than directly by the user.
}
\examples{
library(flowPloidyData) 
fh2 <- FlowHist(file = flowPloidyFiles[12], channel = "FL3.INT.LIN")
plot(fh2, init = TRUE) ## automatic peak estimates
\dontrun{
fh2 <- pickInit(fh2)   ## hand-pick peak estimates
}
plot(fh2, init = TRUE) ## revised starting values
}
\author{
Tyler Smith
}

